Added additional Server Authorization logic.

+++ server/server.route.js
@@ -34,20 +34,24 @@ let serverRoutes = (function () {
     */
   router.get('/account/current', oauthServer.authorise(), (req, res) => {
     if (req.user !== undefined) {
-      return settingsloader.getUserByName(req.user)
-        .then(function (user) {
-          // gets the Userdata (without Pwd)
-          if (user !== null) {
-            res.status(200).send({
-              "username": user.email,
-              "caninvite": user.caninvite,
-              "isAdmin": user.isAdmin
-            });
-          }
-        })
-        .catch((err) => {
-          res.status(400).send(err.message);
-        })
+      // check if the user is allowed to make this request
+      if (req.user === req.oauth.bearerToken.user) {
+        return settingsloader.getUserByName(req.user)
+          .then(function (user) {
+            // gets the Userdata (without Pwd)
+            if (user !== null) {
+              res.status(200).send({
+                "username": user.email,
+                "caninvite": user.caninvite,
+                "isAdmin": user.isAdmin
+              });
+            }
+          })
+          .catch((err) => {
+            res.status(400).send(err.message);
+          })
+      }
+      res.status(401).send('Access denied!');
     }
     res.status(400).send('User is missing!');
   });
@@ -97,38 +101,42 @@ let serverRoutes = (function () {
     if (req.body.username === undefined) {
       res.status(400).send('Required fields missing!');
     } else {
-      // are you allowed to invite new users ?
-      return settingsloader.getUserByName(req.body.username)
-        .then(function (user) {
-          if (user && user.caninvite) {
-            return settingsloader.getUserByName(req.body.friend);
-          }
-          throw new Error('You are not allowed to invite new Users!');
-        })
-        .then(function (user) {
-          // does the user already exists?
-          if (user === null) {
-            const password = generateRandomPassword();
-            const newuser = { email: req.body.friend, password: passwordHash.generate(password), caninvite: false, isAdmin: false };
+      // check if the user is allowed to make this request
+      if (req.body.username === req.oauth.bearerToken.user) {
+        // are you allowed to invite new users ?
+        return settingsloader.getUserByName(req.body.username)
+          .then(function (user) {
+            if (user && user.caninvite) {
+              return settingsloader.getUserByName(req.body.friend);
+            }
+            throw new Error('You are not allowed to invite new Users!');
+          })
+          .then(function (user) {
+            // does the user already exists?
+            if (user === null) {
+              const password = generateRandomPassword();
+              const newuser = { email: req.body.friend, password: passwordHash.generate(password), caninvite: false, isAdmin: false };
 
-            return settingsloader.addUser(newuser);
-          }
-          throw new Error('User already exists!');
-        })
-        .then(function (user) {
-          // Assign Clock Tile to the new User.
-          return settingsloader.assignTile(user, 'clock');
-        })
-        .then(function (user) {
-          return sendMailer.sendMail(user.email, 'Welcome to Dasho ✔', `<b>Hello User!</b> Welcome to <a href="https://dasho.herokuapp.com">dasho</a>. Please login with your E-mail address.`);
-        })
-        .then(function (info) {
-          console.log(info);
-          res.send(JSON.stringify('User successfully added.'));
-        })
-        .catch(err => {
-          res.status(400).send(err.message);
-        });
+              return settingsloader.addUser(newuser);
+            }
+            throw new Error('User already exists!');
+          })
+          .then(function (user) {
+            // Assign Clock Tile to the new User.
+            return settingsloader.assignTile(user, 'clock');
+          })
+          .then(function (user) {
+            return sendMailer.sendMail(user.email, 'Welcome to Dasho ✔', `<b>Hello User!</b> Welcome to <a href="https://dasho.herokuapp.com">dasho</a>. Please login with your E-mail address.`);
+          })
+          .then(function (info) {
+            console.log(info);
+            res.send(JSON.stringify('User successfully added.'));
+          })
+          .catch(err => {
+            res.status(400).send(err.message);
+          });
+      }
+      res.status(401).send('Access denied!');
     }
   });
 
@@ -141,26 +149,30 @@ let serverRoutes = (function () {
   router.put('/changepassword', oauthServer.authorise(), (req, res) => {
     if (req.body.username !== undefined && req.body.password !== undefined
       && req.body.newpassword !== undefined && req.body.newpasswordconfirm !== undefined) {
-      return settingsloader.getUserByName(req.body.username)
-        .then(function (user) {
-          if (user) {
-            if (req.body.newpassword === req.body.newpasswordconfirm) {
-              const checkPassword = passwordHash.verify(req.body.password, user.password);
-              if (checkPassword) {
-                return settingsloader.setPassword(user, req.body.newpassword);
+      // check if the user is allowed to make this request
+      if (req.body.username === req.oauth.bearerToken.user) {
+        return settingsloader.getUserByName(req.body.username)
+          .then(function (user) {
+            if (user) {
+              if (req.body.newpassword === req.body.newpasswordconfirm) {
+                const checkPassword = passwordHash.verify(req.body.password, user.password);
+                if (checkPassword) {
+                  return settingsloader.setPassword(user, req.body.newpassword);
+                }
+                throw new Error('Password is wrong!');
               }
-              throw new Error('Password is wrong!');
+              throw new Error('Password and Confirm Password are not equal!');
             }
-            throw new Error('Password and Confirm Password are not equal!');
-          }
-          throw new Error('Unknown User!');
-        })
-        .then(function (user) {
-          res.send(user);
-        })
-        .catch(err => {
-          res.status(400).send(err.message);
-        });
+            throw new Error('Unknown User!');
+          })
+          .then(function (user) {
+            res.send(user);
+          })
+          .catch(err => {
+            res.status(400).send(err.message);
+          });
+      }
+      res.status(401).send('Access denied!');
     } else {
       res.status(400).send('Required fields missing!');
     }
@@ -207,13 +219,17 @@ let serverRoutes = (function () {
    */
   router.put('/settings/:username', oauthServer.authorise(), (req, res) => {
     if (req.params.username !== undefined) {
-      let setting = req.body.setting;
-      settingsloader.saveSetting(setting)
-        .then(function () {
-          res.status(200).send(true);
-        }).catch(err => {
-          res.status(400).send(err.message);
-        });
+      // check if the user is allowed to make this request
+      if (req.params.username === req.oauth.bearerToken.user) {
+        let setting = req.body.setting;
+        settingsloader.saveSetting(setting)
+          .then(function () {
+            res.status(200).send(true);
+          }).catch(err => {
+            res.status(400).send(err.message);
+          });
+      }
+      res.status(401).send('Access denied!');
     } else {
       res.status(400).send('Required fields missing!');
     }
@@ -227,16 +243,20 @@ let serverRoutes = (function () {
    */
   router.post('/settings/:username/:tile', oauthServer.authorise(), (req, res) => {
     if (req.params.username !== undefined && req.params.tile) {
-      return settingsloader.getUserByName(req.params.username)
-        .then(function (user) {
-          return settingsloader.assignTile(user._id, req.params.tile);
-        })
-        .then(function () {
-          res.send(true);
-        })
-        .catch(err => {
-          res.status(400).send(err.message);
-        });
+      // check if the user is allowed to make this request
+      if (req.params.username === req.oauth.bearerToken.user) {
+        return settingsloader.getUserByName(req.params.username)
+          .then(function (user) {
+            return settingsloader.assignTile(user._id, req.params.tile);
+          })
+          .then(function () {
+            res.send(true);
+          })
+          .catch(err => {
+            res.status(400).send(err.message);
+          });
+      }
+      res.status(401).send('Access denied!');
     }
     res.status(400).send('Required fields missing!');
   });
@@ -249,12 +269,16 @@ let serverRoutes = (function () {
    */
   router.delete('/settings/:username/:id', oauthServer.authorise(), (req, res) => {
     if (req.params.username !== undefined && req.params.id) {
-      return settingsloader.deleteSetting(req.params.id)
-        .then(function () {
-          res.send(true);
-        }).catch(err => {
-          res.status(400).send(err.message);
-        });
+      // check if the user is allowed to make this request
+      if (req.params.username === req.oauth.bearerToken.user) {
+        return settingsloader.deleteSetting(req.params.id)
+          .then(function () {
+            res.send(true);
+          }).catch(err => {
+            res.status(400).send(err.message);
+          });
+      }
+      res.status(401).send('Access denied!');
     }
     res.status(400).send('Required fields missing!');
   });
@@ -267,21 +291,25 @@ let serverRoutes = (function () {
    */
   router.get('/settings/:username', oauthServer.authorise(), (req, res) => {
     if (req.params.username !== undefined) {
-      // check if the user exists
-      return settingsloader.getUserByName(req.params.username)
-        .then(function (user) {
-          if (user) {
-            // returns the list of config
-            return settingsloader.getSettings(user);
-          }
-          throw new Error('Unknown User!');
-        })
-        .then(function (settings) {
-          res.send(settings);
-        })
-        .catch(err => {
-          res.status(400).send(err.message);
-        });
+      // check if the user is allowed to make this request
+      if (req.params.username === req.oauth.bearerToken.user) {
+        // check if the user exists
+        return settingsloader.getUserByName(req.params.username)
+          .then(function (user) {
+            if (user) {
+              // returns the list of config
+              return settingsloader.getSettings(user);
+            }
+            throw new Error('Unknown User!');
+          })
+          .then(function (settings) {
+            res.send(settings);
+          })
+          .catch(err => {
+            res.status(400).send(err.message);
+          });
+      }
+      res.status(401).send('Access denied!');
     }
     res.status(400).send('Required fields missing!');
   });
@@ -294,34 +322,38 @@ let serverRoutes = (function () {
     */
   router.get('/settings/unassigned/:username', oauthServer.authorise(), (req, res) => {
     if (req.params.username !== undefined) {
-      // check if the user exists
-      return settingsloader.getUserByName(req.params.username)
-        .then(function (user) {
-          if (user) {
-            // returns the list of config
-            return settingsloader.getSettings(user);
-          }
-          throw new Error('Unknown User!');
-        })
-        .then(function (settings) {
-          return settingsloader.getTiles(settings).then(function (tiles) {
-            var tilesConfigs = [];
-
-            for (let tileIndex = 0; tileIndex < tiles.length; tileIndex++) {
-              const tile = tiles[tileIndex];
-              // gets all Tiles / compare it with selected Tiles
-              if (settings.find(x => x.tile === tile.name)) {
-                // do nothing
-              } else {
-                tilesConfigs.push(tile);
-              }
+      // check if the user is allowed to make this request
+      if (req.params.username === req.oauth.bearerToken.user) {
+        // check if the user exists
+        return settingsloader.getUserByName(req.params.username)
+          .then(function (user) {
+            if (user) {
+              // returns the list of config
+              return settingsloader.getSettings(user);
             }
-            res.send(tilesConfigs);
+            throw new Error('Unknown User!');
           })
-        })
-        .catch(err => {
-          res.status(400).send(err.message);
-        });
+          .then(function (settings) {
+            return settingsloader.getTiles(settings).then(function (tiles) {
+              var tilesConfigs = [];
+
+              for (let tileIndex = 0; tileIndex < tiles.length; tileIndex++) {
+                const tile = tiles[tileIndex];
+                // gets all Tiles / compare it with selected Tiles
+                if (settings.find(x => x.tile === tile.name)) {
+                  // do nothing
+                } else {
+                  tilesConfigs.push(tile);
+                }
+              }
+              res.send(tilesConfigs);
+            })
+          })
+          .catch(err => {
+            res.status(400).send(err.message);
+          });
+      }
+      res.status(401).send('Access denied!');
     }
     res.status(400).send('Required fields missing!');
   });
